requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
cryptography==37.0.4
h11==0.13.0
idna==3.3
iniconfig==1.1.1
outcome==1.2.0
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycparser==2.21
PyHamcrest==2.0.3
pyOpenSSL==22.0.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==6.2.5
pytest-base-url==1.4.2
pytest-html==3.1.1
pytest-metadata==2.0.2
pytest-selenium==3.0.0
pytest-variables==1.9.0
requests==2.28.1
selenium==3.141.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
tenacity==6.3.1
toml==0.10.2
tomli==2.0.1
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.10
wsproto==1.1.0
